visual updates in change log
